requirements.txt
Flask==1.1.1
Flask-RESTful==0.3.7
boto3
Pillow